scheme client tests
