+++ ydb/public/sdk/cpp/client/ut/helpers/grpc_services/scheme.cpp
@@ -1,0 +1,22 @@
+#include "scheme.h"
+
+namespace NYdb::NScheme {
+
+grpc::Status TSchemeDummyService::DescribePath(
+    grpc::ServerContext* context,
+    const Ydb::Scheme::DescribePathRequest* request,
+    Ydb::Scheme::DescribePathResponse* response
+) {
+    Y_UNUSED(context);
+    Y_UNUSED(request);
+
+    auto* op = response->mutable_operation();
+    op->set_ready(true);
+    op->set_status(Ydb::StatusIds::SUCCESS);
+
+    op->mutable_result()->PackFrom(DummyPathDescription);
+
+    return grpc::Status::OK;
+}
+
+}
