+++ ydb/public/sdk/cpp/client/ut/helpers/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+PEERDIR(
+    ydb/public/api/grpc
+    ydb/public/api/grpc/draft
+)
+
+SRCS(
+    grpc_services/scripting.cpp
+    grpc_services/scheme.cpp
+)
+
+END()
