+++ ydb/public/sdk/cpp/client/ydb_scheme/ut/ya.make
@@ -1,0 +1,13 @@
+UNITTEST_FOR(ydb/public/sdk/cpp/client/ydb_scheme)
+
+SIZE(SMALL)
+
+PEERDIR(
+    ydb/public/sdk/cpp/client/ut/helpers
+)
+
+SRCS(
+    ydb_scheme_ut.cpp
+)
+
+END()
