+++ ydb/public/sdk/cpp/client/ydb_scheme/ya.make
@@ -16,3 +16,7 @@ PEERDIR(
 )
 
 END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
